Cleaned up model definition files.
