+++ opcpublisher/Models/ConfigurationFileEntryModel.cs
@@ -1,0 +1,88 @@
+﻿using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
+using OpcPublisher.Crypto;
+using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+
+namespace OpcPublisher
+{
+    /// <summary>
+    /// Class describing the nodes which should be published.
+    /// </summary>
+    public partial class ConfigurationFileEntryModel
+    {
+        public ConfigurationFileEntryModel()
+        {
+        }
+
+        public ConfigurationFileEntryModel(string endpointUrl)
+        {
+            EndpointUrl = new Uri(endpointUrl);
+            OpcNodes = new List<OpcNodeOnEndpointModel>();
+        }
+
+        public Uri EndpointUrl { get; set; }
+
+        [DefaultValue(true)]
+        [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore, NullValueHandling = NullValueHandling.Ignore)]
+        public bool? UseSecurity { get; set; }
+
+        /// <summary>
+        /// Gets ot sets the authentication mode to authenticate against the OPC UA Server.
+        /// </summary>
+        [DefaultValue(OpcAuthenticationMode.Anonymous)]
+        [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore, NullValueHandling = NullValueHandling.Ignore)]
+        [JsonConverter(typeof(StringEnumConverter))]
+        public OpcAuthenticationMode OpcAuthenticationMode { get; set; }
+
+        /// <summary>
+        /// Gets or sets the encrypted username to authenticate against the OPC UA Server (when OpcAuthenticationMode is set to UsernamePassword
+        /// </summary>
+        [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore, NullValueHandling = NullValueHandling.Ignore)]
+        public string EncryptedAuthUsername
+        {
+            get
+            {
+                return EncryptedAuthCredential?.UserName;
+            }
+            set
+            {
+                if (EncryptedAuthCredential == null)
+                    EncryptedAuthCredential = new EncryptedNetworkCredential();
+
+                EncryptedAuthCredential.UserName = value;
+            }
+        }
+
+        /// <summary>
+        /// Gets or sets the encrypted password to authenticate against the OPC UA Server (when OpcAuthenticationMode is set to UsernamePassword
+        /// </summary>
+        [JsonProperty(DefaultValueHandling = DefaultValueHandling.Ignore, NullValueHandling = NullValueHandling.Ignore)]
+        public string EncryptedAuthPassword
+        {
+            get
+            {
+                return EncryptedAuthCredential?.Password;
+            }
+            set
+            {
+                if (EncryptedAuthCredential == null)
+                    EncryptedAuthCredential = new EncryptedNetworkCredential();
+
+                EncryptedAuthCredential.Password = value;
+            }
+        }
+
+        /// <summary>
+        /// Gets or sets the encrpyted credential to authenticate against the OPC UA Server (when OpcAuthenticationMode is set to UsernamePassword.
+        /// </summary>
+        [JsonIgnore]
+        public EncryptedNetworkCredential EncryptedAuthCredential { get; set; }
+
+        [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
+#pragma warning disable CA2227 // Collection properties should be read only
+        public List<OpcNodeOnEndpointModel> OpcNodes { get; set; }
+#pragma warning restore CA2227 // Collection properties should be read only
+    }
+}
