+++ opcpublisher/Models/DiagnosticLogMethodResponseModel.cs
@@ -1,0 +1,24 @@
+﻿using Newtonsoft.Json;
+using System.Collections.Generic;
+
+namespace OpcPublisher
+{
+    /// <summary>
+    /// Model for a diagnostic log response.
+    /// </summary>
+    public class DiagnosticLogMethodResponseModel
+    {
+        public DiagnosticLogMethodResponseModel()
+        {
+        }
+
+        [JsonProperty(NullValueHandling = NullValueHandling.Include)]
+        public int MissedMessageCount { get; set; }
+
+        [JsonProperty(NullValueHandling = NullValueHandling.Include)]
+        public int LogMessageCount { get; set; }
+
+        [JsonProperty(NullValueHandling = NullValueHandling.Include)]
+        public List<string> Log { get; } = new List<string>();
+    }
+}
