+++ opcpublisher/Models/GetConfiguredEndpointsMethodRequestModel.cs
@@ -1,0 +1,18 @@
+﻿using Newtonsoft.Json;
+
+namespace OpcPublisher
+{
+    /// <summary>
+    /// Model for a get configured endpoints request.
+    /// </summary>
+    public class GetConfiguredEndpointsMethodRequestModel
+    {
+        public GetConfiguredEndpointsMethodRequestModel(ulong? continuationToken = null)
+        {
+            ContinuationToken = continuationToken;
+        }
+
+        [JsonProperty(NullValueHandling = NullValueHandling.Include)]
+        public ulong? ContinuationToken { get; set; }
+    }
+}
