+++ opcpublisher/Models/UnpublishAllNodesMethodRequestModel.cs
@@ -1,0 +1,15 @@
+﻿namespace OpcPublisher
+{
+    /// <summary>
+    /// Model for an unpublish all nodes request.
+    /// </summary>
+    public class UnpublishAllNodesMethodRequestModel
+    {
+        public UnpublishAllNodesMethodRequestModel(string endpointUrl = null)
+        {
+            EndpointUrl = endpointUrl;
+        }
+
+        public string EndpointUrl { get; set; }
+    }
+}
