+++ opcpublisher/Models/UnpublishNodesMethodRequestModel.cs
@@ -1,0 +1,20 @@
+﻿using System.Collections.Generic;
+
+namespace OpcPublisher
+{
+    /// <summary>
+    /// Model for an unpublish node request.
+    /// </summary>
+    public class UnpublishNodesMethodRequestModel
+    {
+        public UnpublishNodesMethodRequestModel(string endpointUrl)
+        {
+            OpcNodes = new List<OpcNodeOnEndpointModel>();
+            EndpointUrl = endpointUrl;
+        }
+
+        public string EndpointUrl { get; set; }
+
+        public List<OpcNodeOnEndpointModel> OpcNodes { get; }
+    }
+}
